+++ apps/web/src/features/scheduled-event/components/scheduled-event-list.tsx
@@ -1,9 +1,9 @@
 import { ParticipantAvatarBar } from "@/components/participant-avatar-bar";
 import { StackedList, StackedListItem } from "@/components/stacked-list";
 import { Trans } from "@/components/trans";
-import { FormattedDateTime } from "@/features/timezone/formatted-date-time";
 import { ScheduledEventStatusBadge } from "@/features/scheduled-event/components/scheduled-event-status-badge";
 import type { Status } from "@/features/scheduled-event/schema";
+import { FormattedDateTime } from "@/features/timezone/client/formatted-date-time";
 
 export const ScheduledEventList = StackedList;
 
@@ -40,7 +40,11 @@ export function ScheduledEventListItem({
             <Trans i18nKey="allDay" defaults="All day" />
           ) : (
             <div className="flex items-center gap-x-1">
-              <FormattedDateTime date={start} floating={isFloating} format="LT" />
+              <FormattedDateTime
+                date={start}
+                floating={isFloating}
+                format="LT"
+              />
               <span>-</span>
               <FormattedDateTime date={end} floating={isFloating} format="LT" />
             </div>
