+++ apps/web/src/features/timezone/index.ts
@@ -1,3 +1,3 @@
-export * from "./timezone-context";
+export * from "./client/context";
 export * from "./timezone-display";
 export * from "./timezone-utils";
